Edit register page to only have username, email, password
