+++ src/environments/environment.staging.ts
@@ -1,4 +1,4 @@
 export const environment = {
   production: true,
-  apiUrl: 'http://localhost:5000/',
+  apiUrl: 'https://staging.kampong.app/',
 };
